requirements.txt
aiosqlite==0.17.0
annotated-types==0.6.0
anyio==4.2.0
click==8.1.7
colorama==0.4.6
fastapi==0.108.0
h11==0.14.0
httptools==0.6.1
idna==3.6
iso8601==1.1.0
pydantic==2.5.3
pydantic_core==2.14.6
pypika-tortoise==0.1.6
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
sniffio==1.3.0
starlette==0.32.0.post1
tortoise-orm==0.20.0
typing_extensions==4.9.0
uvicorn==0.25.0
watchfiles==0.21.0
websockets==12.0
